refactor: move cert out
