+++ core/builder/cipher_builder.go
@@ -3,17 +3,17 @@ package builder
 import (
 	"crypto/x509"
 
+	"github.com/wechatpay-apiv3/wechatpay-go/core"
 	"github.com/wechatpay-apiv3/wechatpay-go/core/cert"
-	"github.com/wechatpay-apiv3/wechatpay-go/core/cert/downloader"
 	"github.com/wechatpay-apiv3/wechatpay-go/core/cipher/encryptors"
 )
 
 func BuildWechatPayEncryptorWithCertificateList(l []*x509.Certificate) *encryptors.WechatPayEncryptor {
-	return encryptors.NewWechatPayEncryptor(cert.NewCertificateMapWithList(l))
+	return encryptors.NewWechatPayEncryptor(core.NewCertificateMapWithList(l))
 }
 
 func BuildWechatPayEncryptorWithDownloaderMgr(
-	mgr *downloader.CertificateDownloaderMgr, mchID string,
+	mgr *cert.CertificateDownloaderMgr, mchID string,
 ) *encryptors.WechatPayEncryptor {
 	return encryptors.NewWechatPayEncryptor(mgr.GetCertificateVisitor(mchID))
 }
