+++ core/cert/downloader_mgr_singleton.go
@@ -1,4 +1,4 @@
-package downloader
+package cert
 
 import (
 	"context"
