+++ core/cert/downloader_test.go
@@ -1,25 +1,20 @@
-package downloader_test
+package cert_test
 
 import (
 	"context"
 	"testing"
-	"time"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/wechatpay-apiv3/wechatpay-go/core"
-	"github.com/wechatpay-apiv3/wechatpay-go/core/cert/downloader"
+	"github.com/wechatpay-apiv3/wechatpay-go/core/cert"
 	"github.com/wechatpay-apiv3/wechatpay-go/core/consts"
 	"github.com/wechatpay-apiv3/wechatpay-go/core/option"
 	"github.com/wechatpay-apiv3/wechatpay-go/utils"
 )
 
-func TestAutoCertificateDownloader(t *testing.T) {
+func TestNewCertificateDownloaderWithClient(t *testing.T) {
 	ctx := context.Background()
-
-	mgr := downloader.NewCertificateDownloaderMgrWithInterval(ctx, 5*time.Second)
-	require.NotNil(t, mgr)
-	defer mgr.Stop()
 
 	privateKey, err := utils.LoadPrivateKey(consts.MchPrivateKey)
 	require.NoError(t, err)
@@ -31,21 +26,28 @@ func TestAutoCertificateDownloader(t *testing.T) {
 	client, err := core.NewClient(ctx, opts...)
 	require.NoError(t, err)
 
-	err = mgr.RegisterDownloaderWithClient(ctx, client, consts.MchID, consts.MchAPIv3Key)
+	d, err := cert.NewCertificateDownloaderWithClient(ctx, client, consts.MchAPIv3Key)
 	require.NoError(t, err)
 
-	err = mgr.RegisterDownloaderWithPrivateKey(ctx, privateKey, consts.SerialNo, consts.MchID, consts.MchAPIv3Key)
-	require.NoError(t, err)
-
-	provider := mgr.GetCertificateVisitor(consts.MchID)
-
-	assert.NotEmpty(t, provider.GetAll(ctx))
-	for serialNo, cert := range provider.GetAll(ctx) {
+	assert.NotEmpty(t, d.GetAll(ctx))
+	for serialNo, cert := range d.GetAll(ctx) {
 		assert.Equal(t, serialNo, utils.GetCertificateSerialNumber(*cert))
 	}
+}
 
-	time.Sleep(11 * time.Second)
+func TestNewCertificateDownloader(t *testing.T) {
+	privateKey, err := utils.LoadPrivateKey(consts.MchPrivateKey)
+	require.NoError(t, err)
 
-	mgr.RemoveDownloader(ctx, consts.MchID)
-	assert.Empty(t, provider.GetAll(ctx))
+	ctx := context.Background()
+
+	d, err := cert.NewCertificateDownloader(
+		context.Background(), consts.MchID, privateKey, consts.SerialNo, consts.MchAPIv3Key,
+	)
+	require.NoError(t, err)
+
+	assert.NotEmpty(t, d.GetAll(ctx))
+	for serialNo, cert := range d.GetAll(ctx) {
+		assert.Equal(t, serialNo, utils.GetCertificateSerialNumber(*cert))
+	}
 }
