+++ core/certificate_map.go
@@ -1,4 +1,4 @@
-package cert
+package core
 
 import (
 	"context"
