+++ core/certificate_visitor.go
@@ -1,4 +1,4 @@
-package cert
+package core
 
 import (
 	"context"
